Update extension boilerplate with solidus_dev_support
