--- .env.example
@@ -1,20 +1,0 @@
-KLARNA_US_API_KEY='xyz'
-KLARNA_US_API_SECRET='xyz'
-
-KLARNA_UK_API_KEY='xyz'
-KLARNA_UK_API_SECRET='xyz'
-
-KLARNA_DE_API_KEY='xyz'
-KLARNA_DE_API_SECRET='xyz'
-
-KLARNA_SE_API_KEY='xyz'
-KLARNA_SE_API_SECRET='xyz'
-
-KLARNA_NO_API_KEY='xyz'
-KLARNA_NO_API_SECRET='xyz'
-
-KLARNA_FI_API_KEY='xyz'
-KLARNA_FI_API_SECRET='xyz'
-
-SOLIDUS_VERSION='~> 1.3.0'
-RAILS_VERSION='~> 4.2.0'
