+++ .rspec
@@ -1,2 +1,2 @@
---format documentation
 --color
+--require spec_helper
