+++ .rubocop.yml
@@ -1,0 +1,9 @@
+inherit_from: .rubocop_todo.yml
+
+require:
+  - solidus_dev_support/rubocop
+
+AllCops:
+  Exclude:
+    - sandbox/**/*
+    - spec/dummy/**/*
