+++ .rubocop_todo.yml
@@ -1,0 +1,215 @@
+# This configuration was generated by
+# `rubocop --auto-gen-config`
+# on 2020-06-11 13:58:01 +0200 using RuboCop version 0.76.0.
+# The point is for the user to remove these configuration records
+# one by one as the offenses are removed from the code base.
+# Note that changes in the inspected code, or installation of new
+# versions of RuboCop, may require this file to be generated again.
+
+# Offense count: 2
+# Cop supports --auto-correct.
+# Configuration parameters: EnforcedStyleAlignWith, AutoCorrect, Severity.
+# SupportedStylesAlignWith: keyword, variable, start_of_line
+Layout/EndAlignment:
+  Exclude:
+    - 'app/serializers/solidus_klarna_payments/line_item_serializer.rb'
+    - 'app/serializers/solidus_klarna_payments/order_serializer.rb'
+
+# Offense count: 2
+Lint/AmbiguousBlockAssociation:
+  Exclude:
+    - 'spec/controllers/spree/klarna/callbacks_controller_spec.rb'
+    - 'spec/spec_helper.rb'
+
+# Offense count: 1
+Lint/DuplicateCaseCondition:
+  Exclude:
+    - 'app/models/solidus_klarna_payments/amount_calculators/uk/order_calculator.rb'
+
+# Offense count: 1
+Lint/DuplicateMethods:
+  Exclude:
+    - 'app/models/spree/klarna_credit_payment.rb'
+
+# Offense count: 1
+# Cop supports --auto-correct.
+Lint/RedundantCopDisableDirective:
+  Exclude:
+    - 'lib/generators/solidus_klarna_payments/install/install_generator.rb'
+
+# Offense count: 3
+Lint/UselessAssignment:
+  Exclude:
+    - 'app/serializers/solidus_klarna_payments/line_item_serializer.rb'
+    - 'spec/serializers/solidus_klarna_payments/order_serializer_spec.rb'
+    - 'spec/support/test_helpers/site_prism_helper.rb'
+
+# Offense count: 1
+# Configuration parameters: CheckForMethodsWithNoSideEffects.
+Lint/Void:
+  Exclude:
+    - 'lib/active_merchant/billing/klarna_gateway.rb'
+
+# Offense count: 1
+Naming/AccessorMethodName:
+  Exclude:
+    - 'spec/support/page_drivers/complete.rb'
+
+# Offense count: 3
+# Configuration parameters: NamePrefix, NamePrefixBlacklist, NameWhitelist, MethodDefinitionMacros.
+# NamePrefix: is_, has_, have_
+# NamePrefixBlacklist: is_, has_, have_
+# NameWhitelist: is_a?
+# MethodDefinitionMacros: define_method, define_singleton_method
+Naming/PredicateName:
+  Exclude:
+    - 'spec/**/*'
+    - 'app/decorators/models/solidus_klarna_payments/spree/payment_decorator.rb'
+    - 'app/models/spree/klarna_credit_payment.rb'
+
+# Offense count: 1
+# Configuration parameters: MinNameLength, AllowNamesEndingInNumbers, AllowedNames, ForbiddenNames.
+# AllowedNames: io, id, to, by, on, in, at, ip, db, os
+Naming/UncommunicativeMethodParamName:
+  Exclude:
+    - 'spec/support/test_helpers/site_prism_helper.rb'
+
+# Offense count: 6
+RSpec/AnyInstance:
+  Exclude:
+    - 'spec/features/checkout/error_handling_spec.rb'
+    - 'spec/serializers/solidus_klarna_payments/discount_serializer_spec.rb'
+    - 'spec/serializers/solidus_klarna_payments/order_serializer_spec.rb'
+    - 'spec/support/test_helpers/checkout_helper.rb'
+
+# Offense count: 19
+# Configuration parameters: Prefixes.
+# Prefixes: when, with, without
+RSpec/ContextWording:
+  Exclude:
+    - 'spec/lib/active_merchant/billing/klarna_gateway_spec.rb'
+    - 'spec/models/spree/payment_spec.rb'
+    - 'spec/serializers/solidus_klarna_payments/address_serializer_spec.rb'
+    - 'spec/serializers/solidus_klarna_payments/discount_serializer_spec.rb'
+    - 'spec/serializers/solidus_klarna_payments/line_item_serializer_spec.rb'
+    - 'spec/serializers/solidus_klarna_payments/order_serializer_spec.rb'
+    - 'spec/support/shared_contexts/klarna_api_helper.rb'
+
+# Offense count: 4
+RSpec/DescribeClass:
+  Exclude:
+    - 'spec/features/admin/payment_management_spec.rb'
+    - 'spec/features/checkout/error_handling_spec.rb'
+    - 'spec/features/checkout/ordering_with_klarna_spec.rb'
+    - 'spec/features/checkout/ordering_without_klarna_spec.rb'
+
+# Offense count: 1
+RSpec/ExpectInHook:
+  Exclude:
+    - 'spec/lib/active_merchant/billing/klarna_gateway_spec.rb'
+
+# Offense count: 1
+# Configuration parameters: CustomTransform, IgnoreMethods.
+RSpec/FilePath:
+  Exclude:
+    - 'spec/serializers/solidus_klarna_payments/discount_serializer_spec.rb'
+
+# Offense count: 14
+RSpec/LetSetup:
+  Exclude:
+    - 'spec/controllers/spree/klarna/sessions_controller_spec.rb'
+    - 'spec/serializers/solidus_klarna_payments/order_serializer_spec.rb'
+
+# Offense count: 19
+# Configuration parameters: .
+# SupportedStyles: have_received, receive
+RSpec/MessageSpies:
+  EnforcedStyle: receive
+
+# Offense count: 36
+RSpec/MultipleExpectations:
+  Max: 6
+
+# Offense count: 4
+# Configuration parameters: IgnoreSharedExamples.
+RSpec/NamedSubject:
+  Exclude:
+    - 'spec/models/spree/gateway/klarna_credit_spec.rb'
+    - 'spec/serializers/solidus_klarna_payments/address_serializer_spec.rb'
+
+# Offense count: 6
+RSpec/NestedGroups:
+  Max: 4
+
+# Offense count: 2
+RSpec/OverwritingSetup:
+  Exclude:
+    - 'spec/serializers/solidus_klarna_payments/order_serializer_spec.rb'
+
+# Offense count: 1
+RSpec/SubjectStub:
+  Exclude:
+    - 'spec/models/spree/gateway/klarna_credit_spec.rb'
+
+# Offense count: 11
+# Configuration parameters: IgnoreNameless, IgnoreSymbolicNames.
+RSpec/VerifiedDoubles:
+  Exclude:
+    - 'spec/lib/active_merchant/billing/klarna_gateway_spec.rb'
+    - 'spec/models/spree/gateway/klarna_credit_spec.rb'
+    - 'spec/support/shared_contexts/klarna_api_helper.rb'
+
+# Offense count: 1
+# Configuration parameters: Include.
+# Include: db/migrate/*.rb
+Rails/CreateTableWithTimestamps:
+  Exclude:
+    - 'db/migrate/20161029131826_create_spree_klarna_credit_payment.rb'
+
+# Offense count: 2
+# Configuration parameters: Include.
+# Include: app/models/**/*.rb
+Rails/InverseOf:
+  Exclude:
+    - 'app/models/spree/klarna_credit_payment.rb'
+
+# Offense count: 1
+Rails/ReflectionClassName:
+  Exclude:
+    - 'app/models/spree/klarna_credit_payment.rb'
+
+# Offense count: 3
+# Configuration parameters: Blacklist, Whitelist.
+# Blacklist: decrement!, decrement_counter, increment!, increment_counter, toggle!, touch, update_all, update_attribute, update_column, update_columns, update_counters
+Rails/SkipsModelValidations:
+  Exclude:
+    - 'app/controllers/spree/klarna/callbacks_controller.rb'
+    - 'lib/solidus_klarna_payments/factories.rb'
+
+# Offense count: 7
+# Configuration parameters: MinBodyLength.
+Style/GuardClause:
+  Exclude:
+    - 'app/controllers/spree/klarna/sessions_controller.rb'
+    - 'app/serializers/solidus_klarna_payments/order_serializer.rb'
+    - 'lib/decorators/backend/solidus_klarna_payments/spree/admin/payment_methods_controller_decorator.rb'
+    - 'lib/decorators/backend/solidus_klarna_payments/spree/admin/payments_controller_decorator.rb'
+    - 'lib/decorators/frontend/solidus_klarna_payments/spree/checkout_controller_decorator.rb'
+    - 'spec/support/test_helpers/site_prism_helper.rb'
+
+# Offense count: 1
+Style/MethodMissingSuper:
+  Exclude:
+    - 'spec/support/test_helpers/site_prism_helper.rb'
+
+# Offense count: 1
+Style/MissingRespondToMissing:
+  Exclude:
+    - 'spec/support/test_helpers/site_prism_helper.rb'
+
+# Offense count: 59
+# Cop supports --auto-correct.
+# Configuration parameters: AutoCorrect, AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
+# URISchemes: http, https
+Metrics/LineLength:
+  Max: 238
