+++ bin/console
@@ -1,14 +1,17 @@
 #!/usr/bin/env ruby
 
+# frozen_string_literal: true
+
 require "bundler/setup"
-require "klarna_gateway"
+require "solidus_klarna_payments"
 
 # You can add fixtures and/or initialization code here to make experimenting
 # with your gem easier. You can also use a different console, if you like.
+$LOAD_PATH.unshift(*Dir["#{__dir__}/../app/*"])
 
 # (If you use this, don't forget to add pry to your Gemfile!)
 # require "pry"
 # Pry.start
 
 require "irb"
-IRB.start
+IRB.start(__FILE__)
