+++ lib/generators/solidus_klarna_payments/templates/initializer.rb
@@ -1,26 +1,6 @@
-module KlarnaGateway
-  module Generators
-    class InstallGenerator < Rails::Generators::Base
+# frozen_string_literal: true
 
-      class_option :auto_run_migrations, :type => :boolean, :default => false
-
-      def add_migrations
-        run 'bundle exec rake railties:install:migrations FROM=klarna_gateway'
-      end
-
-      def add_stylesheets
-        inject_into_file 'vendor/assets/stylesheets/spree/backend/all.css', " *= require spree/backend/klarna_gateway\n", before: /\*\//, verbose: true
-        inject_into_file 'vendor/assets/stylesheets/spree/frontend/all.css', " *= require spree/frontend/klarna_gateway\n", before: / ?\*= ?require_tree \./, verbose: true
-      end
-
-      def add_javascripts
-        inject_into_file 'vendor/assets/javascripts/spree/frontend/all.js', "//= require spree/frontend/klarna_credit\n", before: /\/\/*= ?require_tree \./, verbose: true
-      end
-
-      def add_initializer
-        initializer "klarna_gateway.rb" do
-          <<-INIT
-KlarnaGateway.configure do |config|
+SolidusKlarnaPayments.configure do |config|
   ## Generate the completion route dynamically
   #
   # Expected:
@@ -69,23 +49,8 @@ KlarnaGateway.configure do |config|
   #
   # Example/default:
   #
-  config.product_url = Proc.new do |item|
+  config.product_url = proc do |item|
     host = Spree::Store.default.url.lines.first
     Spree::Core::Engine.routes.url_helpers.product_url(item.variant, host: host) if host.present?
-  end
-end
-          INIT
-        end
-      end
-
-      def run_migrations
-        run_migrations = options[:auto_run_migrations] || ['', 'y', 'Y'].include?(ask 'Would you like to run the migrations now? [Y/n]')
-        if run_migrations
-          run 'bundle exec rake db:migrate'
-        else
-          puts 'Skipping rake db:migrate, don\'t forget to run it!'
-        end
-      end
-    end
   end
 end
