+++ solidus_klarna_payments.gemspec
@@ -1,61 +1,41 @@
-# coding: utf-8
-lib = File.expand_path('../lib', __FILE__)
-$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'klarna_gateway/version'
+# frozen_string_literal: true
+
+require_relative 'lib/solidus_klarna_payments/version'
 
 Gem::Specification.new do |spec|
-  spec.name          = "solidus_klarna_payments"
-  spec.version       = KlarnaGateway::VERSION
-  spec.authors       = ["Jose Antonio Pio Gil", "Pascal Jungblut"]
-  spec.email         = ["jose.pio@bitspire.de", "pascal.jungblut@bitspire.de"]
+  spec.name = 'solidus_klarna_payments'
+  spec.version = SolidusKlarnaPayments::VERSION
+  spec.authors = ['Jose Antonio Pio Gil', 'Pascal Jungblut']
+  spec.email = '["jose.pio@bitspire.de", "pascal.jungblut@bitspire.de"]'
 
-  spec.summary       = %q{Klarna Payments ActiveMerchant gateway and Solidus payment method}
-  spec.description   = %q{Klarna Payments ActiveMerchant gateway and Solidus payment method}
-  spec.homepage      = ""
-  spec.license       = "Apache-2.0"
+  spec.summary = 'Klarna Payments integration for Solidus.'
+  spec.homepage = 'https://github.com/solidusio-contrib/solidus_klarna_payments'
+  spec.license = 'Apache-2.0'
 
-  # Prevent pushing this gem to RubyGems.org. To allow pushes either set the 'allowed_push_host'
-  # to allow pushing to a single host or delete this section to allow pushing to any host.
-  if spec.respond_to?(:metadata)
-    spec.metadata['allowed_push_host'] = "TODO: Set to 'http://mygemserver.com'"
-  else
-    raise "RubyGems 2.0 or newer is required to protect against public gem pushes."
-  end
+  spec.metadata['homepage_uri'] = spec.homepage
+  spec.metadata['source_code_uri'] = 'https://github.com/solidusio-contrib/solidus_klarna_payments'
+  spec.metadata['changelog_uri'] = 'https://github.com/solidusio-contrib/solidus_klarna_payments/blob/master/CHANGELOG.md'
 
-  spec.files         = `git ls-files -z`.split("\x0").reject { |f| f.match(%r{^(test|spec|features)/}) }
-  spec.bindir        = "exe"
-  spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
+  spec.required_ruby_version = Gem::Requirement.new('~> 2.5')
+
+  # Specify which files should be added to the gem when it is released.
+  # The `git ls-files -z` loads the files in the RubyGem that have been added into git.
+  files = Dir.chdir(__dir__) { `git ls-files -z`.split("\x0") }
+
+  spec.files = files.grep_v(%r{^(test|spec|features)/})
+  spec.test_files = files.grep(%r{^(test|spec|features)/})
+  spec.bindir = "exe"
+  spec.executables = files.grep(%r{^exe/}) { |f| File.basename(f) }
   spec.require_paths = ["lib"]
 
-  spec.add_development_dependency "bundler", "~> 1.12"
-  spec.add_development_dependency "rake", "~> 10.0"
+  spec.add_dependency 'activemerchant'
+  spec.add_dependency 'deface'
+  spec.add_dependency 'klarna_client'
+  spec.add_dependency 'solidus_core', ['>= 2.0.0', '< 3']
+  spec.add_dependency 'solidus_support', '~> 0.5'
 
-  spec.add_development_dependency "rspec", "~> 3.0"
-  spec.add_development_dependency "rspec-rails", "~> 3.5.0"
-  spec.add_development_dependency "capybara"
-  spec.add_development_dependency "poltergeist", "~> 1.15.0"
-  spec.add_development_dependency "site_prism"
-  spec.add_development_dependency "vcr", "~> 3.0"
-  spec.add_development_dependency "webmock", "~> 2.0"
-  spec.add_development_dependency "pry-rails"
-  spec.add_development_dependency "awesome_print"
-  spec.add_development_dependency "pry-byebug"
-  spec.add_development_dependency "factory_girl", "~> 4"
-  spec.add_development_dependency "sqlite3"
-  spec.add_development_dependency "database_cleaner"
-  spec.add_development_dependency "pg"
-  spec.add_development_dependency "launchy"
-  spec.add_development_dependency 'selenium-webdriver'
-  spec.add_development_dependency "chromedriver-helper"
-  spec.add_development_dependency "listen"
-  spec.add_development_dependency "dotenv"
-
-  spec.add_dependency "solidus_core", ">= 1.3.0.a", "< 3"
-  spec.add_dependency "solidus_api", ">= 1.3.0.a", "< 3"
-  spec.add_dependency "solidus_frontend", ">= 1.3.0.a", "< 3"
-  spec.add_dependency "solidus_backend", ">= 1.3.0.a", "< 3"
-  spec.add_dependency "solidus_auth_devise"
-
-  spec.add_dependency "klarna_client", ">= 0.9.1"
-  spec.add_dependency "activemerchant"
+  spec.add_development_dependency 'site_prism'
+  spec.add_development_dependency 'solidus_dev_support'
+  spec.add_development_dependency 'vcr'
+  spec.add_development_dependency 'webmock'
 end
